+++ hspec-contrib/test/Test/Hspec/Contrib/HUnitSpec.hs
@@ -1,5 +1,5 @@
 {-# OPTIONS_GHC -fno-warn-orphans #-}
-module Test.Hspec.Contrib.HUnitSpec (main, spec) where
+module Test.Hspec.Contrib.HUnitSpec (spec) where
 
 import           Helper
 
@@ -7,17 +7,8 @@ import           Test.Hspec.Core.Spec
 import           Test.Hspec.Contrib.HUnit
 import           Test.HUnit
 
-main :: IO ()
-main = hspec spec
-
-shouldYield :: Test -> [Tree (ActionWith ()) String] -> Expectation
-a `shouldYield` b = map (Blind . fmap itemRequirement) <$> runSpecM (fromHUnitTest a) `shouldReturn` map Blind b
-
-instance Eq a => Eq (Tree c a) where
-  x == y = case (x, y) of
-    (Node s1 xs, Node s2 ys) -> s1 == s2 && xs == ys
-    (Leaf x1, Leaf x2) -> x1 == x2
-    _ -> False
+shouldYield :: Test -> [Tree () String] -> Expectation
+a `shouldYield` b = map (bimapTree (const ()) itemRequirement) <$> runSpecM (fromHUnitTest a) `shouldReturn` b
 
 spec :: Spec
 spec = do