+++ hspec-core/src/Test/Hspec/Core/Tree.hs
@@ -12,6 +12,7 @@ module Test.Hspec.Core.Tree (
 , Item (..)
 , specGroup
 , specItem
+, bimapTree
 , location
 ) where
 
@@ -28,11 +29,19 @@ data Tree c a =
     Node String [Tree c a]
   | NodeWithCleanup c [Tree c a]
   | Leaf a
-  deriving (Functor, Foldable, Traversable)
+  deriving (Show, Eq, Functor, Foldable, Traversable)
 
 -- | A tree is used to represent a spec internally.  The tree is parametrize
 -- over the type of cleanup actions and the type of the actual spec items.
 type SpecTree a = Tree (ActionWith a) (Item a)
+
+bimapTree :: (a -> b) -> (c -> d) -> Tree a c -> Tree b d
+bimapTree g f = go
+  where
+    go spec = case spec of
+      Node d xs -> Node d (map go xs)
+      NodeWithCleanup cleanup xs -> NodeWithCleanup (g cleanup) (map go xs)
+      Leaf item -> Leaf (f item)
 
 -- |
 -- @Item@ is used to represent spec items internally.  A spec item consists of:
@@ -45,13 +54,20 @@ type SpecTree a = Tree (ActionWith a) (Item a)
 -- example, including QuickCheck properties, Hspec expectations and HUnit
 -- assertions.
 data Item a = Item {
+
   -- | Textual description of behavior
   itemRequirement :: String
+
   -- | Source location of the spec item
 , itemLocation :: Maybe Location
+
   -- | A flag that indicates whether it is safe to evaluate this spec item in
   -- parallel with other spec items
 , itemIsParallelizable :: Maybe Bool
+
+  -- | A flag that indicates whether this spec item is focused.
+, itemIsFocused :: Bool
+
   -- | Example for behavior
 , itemExample :: Params -> (ActionWith a -> IO ()) -> ProgressCallback -> IO Result
 }
@@ -67,7 +83,7 @@ specGroup s = Node msg
 
 -- | The @specItem@ function creates a spec item.
 specItem :: (HasCallStack, Example a) => String -> a -> SpecTree (Arg a)
-specItem s e = Leaf $ Item requirement location Nothing (safeEvaluateExample e)
+specItem s e = Leaf $ Item requirement location Nothing False (safeEvaluateExample e)
   where
     requirement :: HasCallStack => String
     requirement