+++ CHANGES.markdown
@@ -1,6 +1,7 @@
 ## Changes in 2.6.0
-  - Do not calculate diff on `--no-diff` (see #332)
-  - Remove deprecated module Test.Hspec.Core
+   - Allow to focus individual spec items (see #319)
+   - Do not calculate diff on `--no-diff` (see #332)
+   - Remove deprecated module Test.Hspec.Core
 
 ## Changes in 2.5.9
    - Extract source locations from RecConError (see #375)